Refactoring to reuse code that provide authentication and authorization in posts and comments endpoints
